Add some columns from User to RFID

+++ database/migrations/2019_02_19_113732_create_prudcts_discounts_table.php
@@ -17,7 +17,7 @@ class CreatePrudctsDiscountsTable extends Migration
             $table->increments('id');
             $table->integer('rfid_id');
             $table->integer('product_id');
-            $table->integer('discount');
+            $table->double('discount');
             $table->timestamps();
         });
     }
