+++ database/migrations/2019_02_19_113815_create_rfid_limits_table.php
@@ -17,7 +17,7 @@ class CreateRfidLimitsTable extends Migration
             $table->increments('id');
             $table->integer('rfid_id');
             $table->integer('branch_id');
-            $table->integer('limit');
+            $table->double('limit');
             $table->timestamps();
         });
     }
